# backend/api/gen/go/admin/service/v1/i_login_audit_log_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: admin/service/v1/i_login_audit_log.proto

package adminpb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationLoginAuditLogServiceGet = "/admin.service.v1.LoginAuditLogService/Get"
const OperationLoginAuditLogServiceList = "/admin.service.v1.LoginAuditLogService/List"

type LoginAuditLogServiceHTTPServer interface {
	// Get 查询登录审计日志详情
	Get(context.Context, *GetLoginAuditLogRequest) (*LoginAuditLog, error)
	// List 查询登录审计日志列表
	List(context.Context, *v1.PagingRequest) (*ListLoginAuditLogResponse, error)
}

func RegisterLoginAuditLogServiceHTTPServer(s *http.Server, srv LoginAuditLogServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/admin/v1/login-audit-logs", _LoginAuditLogService_List3_HTTP_Handler(srv))
	r.GET("/admin/v1/login-audit-logs/{id}", _LoginAuditLogService_Get3_HTTP_Handler(srv))
}

func _LoginAuditLogService_List3_HTTP_Handler(srv LoginAuditLogServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLoginAuditLogServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListLoginAuditLogResponse)
		return ctx.Result(200, reply)
	}
}

func _LoginAuditLogService_Get3_HTTP_Handler(srv LoginAuditLogServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetLoginAuditLogRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLoginAuditLogServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*GetLoginAuditLogRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LoginAuditLog)
		return ctx.Result(200, reply)
	}
}

type LoginAuditLogServiceHTTPClient interface {
	// Get 查询登录审计日志详情
	Get(ctx context.Context, req *GetLoginAuditLogRequest, opts ...http.CallOption) (rsp *LoginAuditLog, err error)
	// List 查询登录审计日志列表
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *ListLoginAuditLogResponse, err error)
}

type LoginAuditLogServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewLoginAuditLogServiceHTTPClient(client *http.Client) LoginAuditLogServiceHTTPClient {
	return &LoginAuditLogServiceHTTPClientImpl{client}
}

// Get 查询登录审计日志详情
func (c *LoginAuditLogServiceHTTPClientImpl) Get(ctx context.Context, in *GetLoginAuditLogRequest, opts ...http.CallOption) (*LoginAuditLog, error) {
	var out LoginAuditLog
	pattern := "/admin/v1/login-audit-logs/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationLoginAuditLogServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// List 查询登录审计日志列表
func (c *LoginAuditLogServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*ListLoginAuditLogResponse, error) {
	var out ListLoginAuditLogResponse
	pattern := "/admin/v1/login-audit-logs"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationLoginAuditLogServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}