# backend/api/gen/go/admin/service/v1/i_login_policy_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: admin/service/v1/i_login_policy.proto

package adminpb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationLoginPolicyServiceCreate = "/admin.service.v1.LoginPolicyService/Create"
const OperationLoginPolicyServiceDelete = "/admin.service.v1.LoginPolicyService/Delete"
const OperationLoginPolicyServiceGet = "/admin.service.v1.LoginPolicyService/Get"
const OperationLoginPolicyServiceList = "/admin.service.v1.LoginPolicyService/List"
const OperationLoginPolicyServiceUpdate = "/admin.service.v1.LoginPolicyService/Update"

type LoginPolicyServiceHTTPServer interface {
	// Create 创建登录策略
	Create(context.Context, *CreateLoginPolicyRequest) (*emptypb.Empty, error)
	// Delete 删除登录策略
	Delete(context.Context, *DeleteLoginPolicyRequest) (*emptypb.Empty, error)
	// Get 查询登录策略详情
	Get(context.Context, *GetLoginPolicyRequest) (*LoginPolicy, error)
	// List 查询登录策略列表
	List(context.Context, *v1.PagingRequest) (*ListLoginPolicyResponse, error)
	// Update 更新登录策略
	Update(context.Context, *UpdateLoginPolicyRequest) (*emptypb.Empty, error)
}

func RegisterLoginPolicyServiceHTTPServer(s *http.Server, srv LoginPolicyServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/admin/v1/login-policies", _LoginPolicyService_List4_HTTP_Handler(srv))
	r.GET("/admin/v1/login-policies/{id}", _LoginPolicyService_Get4_HTTP_Handler(srv))
	r.POST("/admin/v1/login-policies", _LoginPolicyService_Create3_HTTP_Handler(srv))
	r.PUT("/admin/v1/login-policies/{id}", _LoginPolicyService_Update3_HTTP_Handler(srv))
	r.DELETE("/admin/v1/login-policies/{id}", _LoginPolicyService_Delete3_HTTP_Handler(srv))
}

func _LoginPolicyService_List4_HTTP_Handler(srv LoginPolicyServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLoginPolicyServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListLoginPolicyResponse)
		return ctx.Result(200, reply)
	}
}

func _LoginPolicyService_Get4_HTTP_Handler(srv LoginPolicyServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetLoginPolicyRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLoginPolicyServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*GetLoginPolicyRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LoginPolicy)
		return ctx.Result(200, reply)
	}
}

func _LoginPolicyService_Create3_HTTP_Handler(srv LoginPolicyServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateLoginPolicyRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLoginPolicyServiceCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Create(ctx, req.(*CreateLoginPolicyRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _LoginPolicyService_Update3_HTTP_Handler(srv LoginPolicyServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateLoginPolicyRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLoginPolicyServiceUpdate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Update(ctx, req.(*UpdateLoginPolicyRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _LoginPolicyService_Delete3_HTTP_Handler(srv LoginPolicyServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteLoginPolicyRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationLoginPolicyServiceDelete)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Delete(ctx, req.(*DeleteLoginPolicyRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

type LoginPolicyServiceHTTPClient interface {
	// Create 创建登录策略
	Create(ctx context.Context, req *CreateLoginPolicyRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// Delete 删除登录策略
	Delete(ctx context.Context, req *DeleteLoginPolicyRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// Get 查询登录策略详情
	Get(ctx context.Context, req *GetLoginPolicyRequest, opts ...http.CallOption) (rsp *LoginPolicy, err error)
	// List 查询登录策略列表
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *ListLoginPolicyResponse, err error)
	// Update 更新登录策略
	Update(ctx context.Context, req *UpdateLoginPolicyRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
}

type LoginPolicyServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewLoginPolicyServiceHTTPClient(client *http.Client) LoginPolicyServiceHTTPClient {
	return &LoginPolicyServiceHTTPClientImpl{client}
}

// Create 创建登录策略
func (c *LoginPolicyServiceHTTPClientImpl) Create(ctx context.Context, in *CreateLoginPolicyRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/login-policies"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationLoginPolicyServiceCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Delete 删除登录策略
func (c *LoginPolicyServiceHTTPClientImpl) Delete(ctx context.Context, in *DeleteLoginPolicyRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/login-policies/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationLoginPolicyServiceDelete))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Get 查询登录策略详情
func (c *LoginPolicyServiceHTTPClientImpl) Get(ctx context.Context, in *GetLoginPolicyRequest, opts ...http.CallOption) (*LoginPolicy, error) {
	var out LoginPolicy
	pattern := "/admin/v1/login-policies/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationLoginPolicyServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// List 查询登录策略列表
func (c *LoginPolicyServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*ListLoginPolicyResponse, error) {
	var out ListLoginPolicyResponse
	pattern := "/admin/v1/login-policies"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationLoginPolicyServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Update 更新登录策略
func (c *LoginPolicyServiceHTTPClientImpl) Update(ctx context.Context, in *UpdateLoginPolicyRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/admin/v1/login-policies/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationLoginPolicyServiceUpdate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}